Making plusSlides() fail safe and autonomous.
Changed so that the value is slides.length instead of 6. This means if I add a new image to the sider I do not have to update this code
Added a fail safe with an else statement. If any value falls out of the slides parameters it will have a default value of currentSlide(1)